pyproject.toml
[project]
name = "reasonance"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Orpheus Lummis", email = "o@orpheuslummis.info" }
]
dependencies = [
    "assemblyai[extras]>=0.35.1",
    "python-dotenv>=1.0.1",
    "keyboard>=0.13.5",
    "pynput>=1.7.7",
    "uvicorn>=0.32.1",
    "fastapi>=0.115.5",
    "python-multipart>=0.0.17",
    "requests>=2.32.3",
    "sse-starlette>=2.1.3",
    "anthropic>=0.39.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/reasonance"]
